v8: harden core helpers and raise coverage
Reject malformed, truncated, indefinite, negative, and overflowing ASN.1 length encodings without panics. Add boundary coverage for length encoding and application-tag wrapping, and exercise ccache address and authorization-data decoding through round trips and malformed fields.

Handle invalid host addresses consistently, preserve duplicate-sensitive address equality, and make Kerberos flag access safe for nil and out-of-range inputs. Add focused tests for IPv4, IPv6, local addresses, equality semantics, and flag boundaries.

Use width-safe NEGOEX and PAC message-size arithmetic so framing guards compile and reject oversized payloads on 32-bit targets. Raise cross-package statement coverage from 65.0 to 65.9 percent and ratchet the CI floor to 65.0 percent.

Validated with go test ./..., go vet ./..., GOARCH=386 go test ./..., cross-platform Windows, macOS, Android, and iOS build checks, race tests, Go 1.18 tests, and the atomic cross-package coverage gate.

+++ v8/pac/marshal.go
@@ -103,7 +103,7 @@ func marshalPACBuffers(version uint32, buffers []pacBuffer) ([]byte, error) {
 	headerLength := uint64(8 + len(buffers)*16)
 	offset := alignPAC(headerLength)
 	for _, buffer := range buffers {
-		if len(buffer.data) > math.MaxUint32 || offset > math.MaxUint64-uint64(len(buffer.data))-pacAlignment+1 {
+		if uint64(len(buffer.data)) > math.MaxUint32 || offset > math.MaxUint64-uint64(len(buffer.data))-pacAlignment+1 {
 			return nil, fmt.Errorf("%w: PAC is too large", ErrPACMalformed)
 		}
 		offset = alignPAC(offset + uint64(len(buffer.data)))
